Stop compiler error

+++ test/test_fsw_fault_handler/test_main_fh_end_to_end.cpp
@@ -4,7 +4,7 @@
 ////////////// Helper utilities ////////////////
 
 void signal_until_persistence(std::shared_ptr<Fault>& fault_ptr) {
-    for(int i = 0; i < fault_ptr->persistence_f.get(); i++)
+    for(size_t i = 0; i < fault_ptr->persistence_f.get(); i++)
         fault_ptr->signal();
 }
